# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Test Spot2",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://test-spot-two.loc:31080"
        }
    ],
    "paths": {
        "/api/urls": {
            "get": {
                "summary": "Obtener la lista de URLs acortadas",
                "operationId": "c34aa4174db5480ce711e7055d9aa392",
                "responses": {
                    "200": {
                        "description": "Lista de URLs acortadas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Url lists"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "original_url": {
                                                        "type": "string",
                                                        "example": "https://google.com"
                                                    },
                                                    "shortened_url": {
                                                        "type": "string",
                                                        "example": "http://localhost:8000/abc123"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear URL Shortener",
                "operationId": "9344f793e06af23abec09fa690b8ba14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "https://google.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL acortada generada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url_decode": {
                                            "type": "string",
                                            "example": "https://google.com"
                                        },
                                        "url_encode": {
                                            "type": "string",
                                            "example": "v4c9r9DT"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación de entrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Por favor, ingresa una URL válida."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/urls/{shortenUrl}": {
            "get": {
                "summary": "Mostrar URL original para una URL acortada",
                "operationId": "1e0c0d305dc3d0c2a0e4af647571f96c",
                "parameters": [
                    {
                        "name": "shortenUrl",
                        "in": "path",
                        "description": "El identificador de la URL acortada",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "KVDtivYE"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL original encontrada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La shortenUrl KVDtivYE fué encontrada"
                                        },
                                        "data": {
                                            "properties": {
                                                "original_url": {
                                                    "type": "string",
                                                    "example": "https://google.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La URL acortada no fue encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "NOK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La shortenUrl KVDtivYE no fué encontrada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/shortenUrl/{shortenUrl}": {
            "delete": {
                "summary": "Eliminar una URL acortada",
                "operationId": "b704482e9a8881b1b3a03966eeef5bc1",
                "parameters": [
                    {
                        "name": "shortenUrl",
                        "in": "path",
                        "description": "El identificador de la URL acortada",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL acortada eliminada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The shortenUrl abc123 deleted"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La URL acortada no fue encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "NOK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The shortenUrl abc123 was not found"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}